+++ CHANGELOG.md
@@ -7,6 +7,8 @@ The project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html)
 
 ### Changed
 
+- Update `secrecy` version to 0.10; change `Secret` to `SecretBox` accordingly, and make `Seed` a struct
+  instead of a type alias. 
 - Bump MSRV to 1.70.
 
 ## 0.5.0 - 2022-07-21